dts: rp1: hobble DMA AXI burst lengths
Channels 1-2 have a statically configured maximum MSIZE of 8, and
channels 3-8 have MSIZE set to 4. The DMAC "helpfully" silently
truncates bursts to the hardware supported maximum, so any FIFO read
operation with an oversized burst threshold will leave a residue of
threshold minus MSIZE rows.

As channel allocation is dynamic, this means every client needs to use a
maximum of 4 for burst length.

AXI AWLEN/ARLEN constraints aren't strictly related to MSIZE, except
that bursts won't be issued that are longer than MSIZE beats. Therefore,
it's a useful proxy to tell clients of the DMAC the hardware
limitations.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>